+++ tests/test_notify.py
@@ -8,7 +8,7 @@ def test_encrypt_email_valid():
         "weaver.wps_email_encrypt_salt": "salty-email",
     }
     email = encrypt_email("some@email.com", settings)
-    assert email == u"a1724b030d999322e2ecc658453f992472c63867cd3cef3b3d829d745bd80f34"
+    assert email == "a1724b030d999322e2ecc658453f992472c63867cd3cef3b3d829d745bd80f34"
 
 
 def test_encrypt_email_raise():