+++ weaver/wps_restapi/swagger_definitions.py
@@ -1,7 +1,7 @@
 """
 Schema definitions for `OpenAPI` generation and validation of data from received requests and returned responses.
 
-This module should contain any and every definitions in use to build the Swagger UI and the OpenAPI JSON schema
+This module should contain any and every definition in use to build the Swagger UI and the OpenAPI JSON schema
 so that one can update the specification without touching any other files after the initial integration.
 
 Schemas defined in this module are employed (through ``deserialize`` method calls) to validate that data conforms to
@@ -54,7 +54,9 @@ from weaver.wps_restapi.colander_extras import (
     AllOfKeywordSchema,
     AnyOfKeywordSchema,
     BoundedRange,
+    CommaSeparated,
     EmptyMappingSchema,
+    ExpandStringList,
     ExtendedBoolean as Boolean,
     ExtendedFloat as Float,
     ExtendedInteger as Integer,
@@ -1730,9 +1732,28 @@ class QuoteSortEnum(ExtendedSchemaNode):
     validator = OneOf(SortMethods.QUOTE)
 
 
+class JobTagsCommaSeparated(ExpandStringList, ExtendedSchemaNode):
+    schema_type = String
+    validator = CommaSeparated()
+    default = None
+    missing = drop
+    description = (
+        "Comma-separated tags that can be used to filter jobs. "
+        f"Only {validator.allow_chars} characters are permitted."
+    )
+
+
+class JobGroupsCommaSeparated(ExpandStringList, ExtendedSchemaNode):
+    schema_type = String
+    default = None
+    example = "process,service"
+    missing = drop
+    description = "Comma-separated list of grouping fields with which to list jobs."
+    validator = StringOneOf(["process", "provider", "service", "status"], delimiter=",", case_sensitive=True)
+
+
 class LaunchJobQuerystring(ExtendedMappingSchema):
-    tags = ExtendedSchemaNode(String(), title="JobTags", default=None, missing=drop,
-                              description="Comma separated tags that can be used to filter jobs later")
+    tags = JobTagsCommaSeparated()
 
 
 class VisibilityValue(ExtendedSchemaNode):
@@ -4700,9 +4721,7 @@ class GetJobsQueries(ExtendedMappingSchema):
     #   Items with default value are added if omitted, except 'default=null' which are removed after handling by alias.
     detail = ExtendedSchemaNode(QueryBoolean(), default=False, example=True, missing=drop,
                                 description="Provide job details instead of IDs.")
-    groups = ExtendedSchemaNode(String(),
-                                description="Comma-separated list of grouping fields with which to list jobs.",
-                                default=False, example="process,service", missing=drop)
+    groups = JobGroupsCommaSeparated()
     page = ExtendedSchemaNode(Integer(allow_string=True), missing=0, default=0, validator=Range(min=0))
     limit = ExtendedSchemaNode(Integer(allow_string=True), missing=10, default=10, validator=Range(min=1, max=10000))
     min_duration = ExtendedSchemaNode(
@@ -4717,7 +4736,13 @@ class GetJobsQueries(ExtendedMappingSchema):
     processID = ProcessIdentifierTag(missing=drop, default=null,
                                      description="Alias to 'process' for OGC-API compliance.")
     process = ProcessIdentifierTag(missing=drop, default=None,
-                                   description="Identifier of the process to filter search.")
+                                   description="Identifier and optional version tag of the process to filter search.")
+    version = Version(
+        missing=drop, default=None, example="0.1.0", description=(
+            "Version of the 'process' or 'processID' query parameters. "
+            "If version is provided, those query parameters should specify the ID without tag."
+        )
+    )
     service = AnyIdentifier(missing=drop, default=null, description="Alias to 'provider' for backward compatibility.")
     provider = AnyIdentifier(missing=drop, default=None, description="Identifier of service provider to filter search.")
     type = JobTypeEnum(missing=drop, default=null,
@@ -4725,8 +4750,7 @@ class GetJobsQueries(ExtendedMappingSchema):
     sort = JobSortEnum(missing=drop)
     access = JobAccess(missing=drop, default=None)
     notification_email = ExtendedSchemaNode(String(), missing=drop, validator=Email())
-    tags = ExtendedSchemaNode(String(), missing=drop, default=None,
-                              description="Comma-separated values of tags assigned to jobs.")
+    tags = JobTagsCommaSeparated()
 
 
 class GetProcessJobsQuery(LocalProcessQuery, GetJobsQueries):